2004-12-28  Zoltan Varga  <vargaz@freemail.hu>
	* X11Structs.cs X11Keyboard.cs XplatUIX11.cs: Fix 64 bit issues.

svn path=/trunk/mcs/; revision=38146

+++ mcs/class/Managed.Windows.Forms/System.Windows.Forms/X11Keyboard.cs
@@ -75,7 +75,7 @@ namespace System.Windows.Forms {
 			if ((xevent.KeyEvent.keycode >> 8) == 0x10)
 				xevent.KeyEvent.keycode = xevent.KeyEvent.keycode & 0xFF;
 
-			int event_time = xevent.KeyEvent.time;
+			int event_time = (int)xevent.KeyEvent.time;
 			int vkey = EventToVkey (xevent);
 
 			if (vkey == 0)
